# openpecha_api/api-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenPecha API v2",
    "version": "2.11.0"
  },
  "paths": {
...
    "/v2/texts": {
      "get": {
        "tags": [
          "Texts"
        ],
        "summary": "List all texts",
        "description": "Retrieve a paginated list of texts (texts) with optional filters.",
        "operationId": "get_all_texts_v2_texts_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Language"
            }
          },
          {
            "name": "title",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/models__base__NonEmptyStr__2"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by title, minimum 2 characters",
              "title": "Title"
            },
            "description": "Filter by title, minimum 2 characters"
          },
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category Id"
            }
          },
          {
            "name": "tag_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag Id"
            }
          },
          {
            "name": "author_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Author Id"
            }
          },
          {
            "name": "bdrc",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bdrc"
            }
          },
          {
            "name": "wiki",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Wiki"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum number of items to return",
              "default": 20,
              "title": "Limit"
            },
            "description": "Maximum number of items to return"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip"
          },
          {
            "name": "X-Application",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Application"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_TextOutput_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Texts"
        ],
        "summary": "Create a text",
        "description": "Create a new text (text) with the provided data.",
        "operationId": "create_text_v2_texts_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v2/texts/{text_id}": {
      "get": {
        "tags": [
          "Texts"
        ],
        "summary": "Get a text",
        "description": "Retrieve a single text (text) by its ID.",
        "operationId": "get_text_v2_texts__text_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "text_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The ID of the text to retrieve",
              "title": "Text Id"
            },
            "description": "The ID of the text to retrieve"
          },
          {
            "name": "X-Application",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Application"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextOutput"
                }
              }
            }
          },
          "404": {
            "description": "Text not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Texts"
        ],
        "summary": "Update a text",
        "description": "Partially update a text with the provided data.",
        "operationId": "update_text_v2_texts__text_id__patch",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "text_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The ID of the text to update",
              "title": "Text Id"
            },
            "description": "The ID of the text to update"
          },
          {
            "name": "X-Application",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Application"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Texts"
        ],
        "summary": "Delete text",
        "description": "Delete a text if it has no editions or incoming text relationships.",
        "operationId": "delete_text_v2_texts__text_id__delete",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "text_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The ID of the text to delete",
              "title": "Text Id"
            },
            "description": "The ID of the text to delete"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AIContribution-Input": {
        "properties": {
          "type": {
            "type": "string",
            "const": "ai",
            "title": "Type"
          },
          "id": {
            "$ref": "#/components/schemas/models__base__NonEmptyStr__1"
          },
          "role": {
            "$ref": "#/components/schemas/ContributorRole"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "id",
          "role"
        ],
        "title": "AIContribution"
      },
      "AIContribution-Output": {
        "properties": {
          "type": {
            "type": "string",
            "const": "ai",
            "title": "Type"
          },
          "id": {
            "$ref": "#/components/schemas/NonEmptyStr-Output"
          },
          "role": {
            "$ref": "#/components/schemas/ContributorRole"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "id",
          "role"
        ],
        "title": "AIContribution"
      },
...
      "ContributionInputItem": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/PersonContributionInput"
          },
          {
            "$ref": "#/components/schemas/AIContribution-Input"
          }
        ],
        "discriminator": {
          "propertyName": "type",
          "mapping": {
            "ai": "#/components/schemas/AIContribution-Input",
            "person": "#/components/schemas/PersonContributionInput"
          }
        }
      },
      "ContributionOutputItem": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/PersonContributionOutput"
          },
          {
            "$ref": "#/components/schemas/AIContribution-Output"
          }
        ],
        "discriminator": {
          "propertyName": "type",
          "mapping": {
            "ai": "#/components/schemas/AIContribution-Output",
            "person": "#/components/schemas/PersonContributionOutput"
          }
        }
      },
      "ContributorRole": {
        "type": "string",
        "enum": [
          "translator",
          "reviser",
          "author",
          "scholar",
          "narrator"
        ],
        "title": "ContributorRole"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IdResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id"
        ],
        "title": "IdResponse"
      },
...
      "LicenseType": {
        "type": "string",
        "enum": [
          "cc0",
          "public",
          "cc-by",
          "cc-by-sa",
          "cc-by-nd",
          "cc-by-nc",
          "cc-by-nc-sa",
          "cc-by-nc-nd",
          "copyrighted",
          "unknown"
        ],
        "title": "LicenseType"
      },
      "LocalizedString-Input": {
        "additionalProperties": {
          "$ref": "#/components/schemas/models__base__NonEmptyStr__1"
        },
        "type": "object",
        "minProperties": 1,
        "title": "LocalizedString"
      },
      "LocalizedString-Output": {
        "additionalProperties": {
          "$ref": "#/components/schemas/NonEmptyStr-Output"
        },
        "type": "object",
        "minProperties": 1,
        "title": "LocalizedString"
      },
      "NonEmptyStr-Output": {
        "type": "string",
        "minLength": 1,
        "pattern": "\\S"
      },
...
      "PaginatedResponse_TextOutput_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TextOutput"
            },
            "type": "array",
            "title": "Items"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "has_more",
          "offset",
          "limit"
        ],
        "title": "PaginatedResponse[TextOutput]"
      },
...
      "PersonContributionInput": {
        "properties": {
          "type": {
            "type": "string",
            "const": "person",
            "title": "Type"
          },
          "id": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/models__base__NonEmptyStr__1"
              },
              {
                "type": "null"
              }
            ]
          },
          "bdrc_id": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/models__base__NonEmptyStr__1"
              },
              {
                "type": "null"
              }
            ]
          },
          "role": {
            "$ref": "#/components/schemas/ContributorRole"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "role"
        ],
        "title": "PersonContributionInput"
      },
      "PersonContributionOutput": {
        "properties": {
          "type": {
            "type": "string",
            "const": "person",
            "title": "Type"
          },
          "id": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NonEmptyStr-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "bdrc_id": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NonEmptyStr-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "role": {
            "$ref": "#/components/schemas/ContributorRole"
          },
          "name": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LocalizedString-Output"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "role"
        ],
        "title": "PersonContributionOutput"
      },
...
      "TextInput": {
        "properties": {
          "bdrc": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/models__base__NonEmptyStr__1"
              },
              {
                "type": "null"
              }
            ]
          },
          "wiki": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/models__base__NonEmptyStr__1"
              },
              {
                "type": "null"
              }
            ]
          },
          "date": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/models__base__NonEmptyStr__1"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "$ref": "#/components/schemas/LocalizedString-Input"
          },
          "alt_titles": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/LocalizedString-Input"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alt Titles"
          },
          "language": {
            "$ref": "#/components/schemas/models__base__NonEmptyStr__1"
          },
          "commentary_of": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/models__base__NonEmptyStr__1"
              },
              {
                "type": "null"
              }
            ]
          },
          "translation_of": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/models__base__NonEmptyStr__1"
              },
              {
                "type": "null"
              }
            ]
          },
          "category_id": {
            "$ref": "#/components/schemas/models__base__NonEmptyStr__1"
          },
          "license": {
            "$ref": "#/components/schemas/LicenseType",
            "default": "public"
          },
          "contributions": {
            "items": {
              "$ref": "#/components/schemas/ContributionInputItem"
            },
            "type": "array",
            "title": "Contributions"
          },
          "tag_ids": {
            "items": {
              "$ref": "#/components/schemas/models__base__NonEmptyStr__1"
            },
            "type": "array",
            "title": "Tag Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "language",
          "category_id"
        ],
        "title": "TextInput"
      },
      "TextOutput": {
        "properties": {
          "bdrc": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NonEmptyStr-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "wiki": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NonEmptyStr-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "date": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NonEmptyStr-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "$ref": "#/components/schemas/LocalizedString-Output"
          },
          "alt_titles": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/LocalizedString-Output"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alt Titles"
          },
          "language": {
            "$ref": "#/components/schemas/NonEmptyStr-Output"
          },
          "commentary_of": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NonEmptyStr-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "translation_of": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NonEmptyStr-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "category_id": {
            "$ref": "#/components/schemas/NonEmptyStr-Output"
          },
          "license": {
            "$ref": "#/components/schemas/LicenseType",
            "default": "public"
          },
          "id": {
            "$ref": "#/components/schemas/NonEmptyStr-Output"
          },
          "contributions": {
            "items": {
              "$ref": "#/components/schemas/ContributionOutputItem"
            },
            "type": "array",
            "title": "Contributions"
          },
          "commentaries": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Commentaries"
          },
          "translations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Translations"
          },
          "editions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Editions"
          },
          "tag_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tag Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "language",
          "category_id",
          "id",
          "contributions"
        ],
        "title": "TextOutput"
      },
      "TextPatch": {
        "properties": {
          "bdrc": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/models__base__NonEmptyStr__1"
              },
              {
                "type": "null"
              }
            ]
          },
          "wiki": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/models__base__NonEmptyStr__1"
              },
              {
                "type": "null"
              }
            ]
          },
          "date": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/models__base__NonEmptyStr__1"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LocalizedString-Input"
              },
              {
                "type": "null"
              }
            ]
          },
          "alt_titles": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/LocalizedString-Input"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alt Titles"
          },
          "language": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/models__base__NonEmptyStr__1"
              },
              {
                "type": "null"
              }
            ]
          },
          "category_id": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/models__base__NonEmptyStr__1"
              },
              {
                "type": "null"
              }
            ]
          },
          "license": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LicenseType"
              },
              {
                "type": "null"
              }
            ]
          },
          "contributions": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ContributionInputItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contributions"
          },
          "tag_ids": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/models__base__NonEmptyStr__1"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "TextPatch"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "models__base__NonEmptyStr__1": {
        "type": "string",
        "minLength": 1,
        "pattern": "\\S"
      },
      "models__base__NonEmptyStr__2": {
        "type": "string",
        "minLength": 2,
        "pattern": "\\S"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}